rename c2cri with c2c
