+++ src/main/java/usdot/fhwa/stol/c2c/c2c_mvt/ServletInitializer.java
@@ -1,4 +1,4 @@
-package usdot.fhwa.stol.c2cri.c2cri_mvt;
+package usdot.fhwa.stol.c2c.c2c_mvt;
 
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
@@ -7,7 +7,7 @@ public class ServletInitializer extends SpringBootServletInitializer {
 
 	@Override
 	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(C2criMvtApplication.class);
+		return application.sources(c2cMvtApplication.class);
 	}
 
 }
