+++ src/main/java/usdot/fhwa/stol/c2c/c2c_mvt/c2cMvtApplication.java
@@ -1,13 +1,13 @@
-package usdot.fhwa.stol.c2cri.c2cri_mvt;
+package usdot.fhwa.stol.c2c.c2c_mvt;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class C2criMvtApplication {
+public class c2cMvtApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(C2criMvtApplication.class, args);
+		SpringApplication.run(c2cMvtApplication.class, args);
 	}
 
 }
