+++ src/main/resources/application.properties
@@ -1,2 +1,2 @@
-spring.application.name=c2cri-mvt
-logging.level.usdot.fhwa.stol.c2cri.c2cri_mvt=DEBUG
+spring.application.name=c2c-mvt
+logging.level.usdot.fhwa.stol.c2c.c2c_mvt=DEBUG
